improve mousewheel handling

+++ galaxy/src/galaxy/events/MouseWheel.cpp
@@ -12,11 +12,11 @@ namespace galaxy
 	namespace events
 	{
 		MouseWheel::MouseWheel() noexcept
-		    : m_x_offset {0.0}, m_y_offset {0.0}
+		    : m_x_offset {0}, m_y_offset {0}
 		{
 		}
 
-		MouseWheel::MouseWheel(const double x, const double y) noexcept
+		MouseWheel::MouseWheel(const int x, const int y) noexcept
 		    : m_x_offset {x}, m_y_offset {y}
 		{
 		}
